requirements.txt
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
openpyxl>=3.0.10
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
python-dateutil>=2.8.2
pytz>=2022.7
